pyproject.toml
[project]
name = "medagent"
version = "0.1.0"
description = "MedAgent Project"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "chromadb>=1.3.7",
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-google-vertexai>=3.2.0",
    "langchain-huggingface>=1.2.0",
    "pysqlite3-binary>=0.5.4.post2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.1",
    "torch==2.5.1",
    "torchvision==0.20.1",
]
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
